requirements.txt
tensorflow
matplotlib
numpy
tqdm
seaborn
tensornetwork
pytorch
